apps/api/pyproject.toml
[project]
name = "career-pilot-api"
version = "0.1.0"
description = "FastAPI backend for CareerPilot."
readme = "README.md"
requires-python = "==3.11.15"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.4",
    "minio>=7.2.18",
    "pypdf>=6.1.1",
    "pydantic-settings>=2.11.0",
    "PyJWT>=2.10.1",
    "pwdlib[argon2]>=0.3.0",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
